cleaning up middleware
