+++ resume_api/apis/v1/middleware.py
@@ -1,69 +1,24 @@
-import json
-
 from flask import current_app, request, g
 from werkzeug.wrappers import Response
 from werkzeug.exceptions import HTTPException
+from resume_api.models import User
 
-from .utils import decode_client_token, get_token_from_request
-from resume_api.models import Client
-
-def check_auth(auth):
-    un = current_app.config['BASIC_AUTH_USERNAME']
-    pw = current_app.config['BASIC_AUTH_PASSWORD']
-    return auth.username == un and auth.password == pw
-
-
-def basic_auth(endpoint):
-
+def verify_password(endpoint):
     auth = request.authorization
-    if not auth or not check_auth(auth):
-        resp = Response(
-            response=json.dumps({'message': 'Not Authorized', 'error': True}),
-            headers = {
-                'WWW-Authenticate': 'Basic realm="Login Required"'
-            },
-            mimetype='application/json',
-            status=401
-        )
-        raise HTTPException('Api Error', response=resp)
 
+    if auth is None:
+        payload = { 'message': 'Unauthorized.' }
+        endpoint.return_error(403, payload=payload)
 
-def get_api_client_from_request(endpoint):
-    """Ensure a valid api_key query param was provided with the request and
-    store the client on `g.api_client`
-    """
+    username_or_token = auth.username
+    password = auth.password
 
-    api_key = request.args.get('api_key', None)
-    if not api_key:
-        payload = { 'message': 'Missing required ?api_key param' }
-        endpoint.return_error(422, payload=payload)
-
-    client = Client.query.filter_by(client_id=api_key).one_or_none()
-    if client is None:
-        payload = { 'message': 'API Client not found' }
-        endpoint.return_error(404, payload=payload)
-
-    g.api_client = client
-
-def get_client_token(endpoint):
-    """Validate the request has a valid JWT token. If the provided api_key
-    and token are valid the JWT will be decoded and stored on g.token for
-    use later on.
-    """
-    token = get_token_from_request()
-    if not token:
-        payload = { 'message': 'Request does not contain a token' }
-        endpoint.return_error(
-            401,
-            payload=payload
-        )
-
-    payload = decode_client_token(g.api_client, token)
-    if not payload:
-        payload = { 'message': 'Invalid token' }
-        endpoint.return_error(
-            401,
-            payload=payload
-        )
-
-    g.token = payload
+    # first try to authenticate by token
+    user = User.verify_auth_token(username_or_token)
+    if not user:
+        # try to authenticate with username/password
+        user = User.query.filter_by(username = username_or_token).first()
+        if not user or not user.verify_password(password):
+            return False
+    g.user = user
+    return True
